LayoutTests/fast/dom/Window/window-postmessage-clone-really-deep-array.html

<html>
<head></head>
<body>
<div id="description"></div>
<input type="file" id="fileInput" multiple></input>
<div id="console"></div>
<script src="resources/postmessage-test.js"></script>
<script>
document.getElementById("description").innerHTML = "Tests that we abort cloning overdeep arrays.";
reallyDeepArray=[];
for (var i = 0; i < 100000; i++)
    reallyDeepArray=[reallyDeepArray];
tryPostMessage('reallyDeepArray', true, null, DOMException.DATA_CLONE_ERR);
tryPostMessage('"done"');
</script>
</body>
</html>
