LayoutTests/platform/chromium/fast/dom/Window/window-postmessage-clone-really-deep-array-expected.txt

Tests that we abort cloning overdeep arrays.

PASS: 'postMessage(reallyDeepArray)' threw Error: DATA_CLONE_ERR: DOM Exception 25
PASS: eventData is done of type string
